+++ pccontrol/TextRecognization/OcrDaemonClient.h
@@ -4,6 +4,6 @@
 #import <Foundation/Foundation.h>
 
 // 1 = daemon returned a response, 0 = daemon unavailable, -1 = daemon error.
-int ZXPerformOcrThroughDaemon(UInt8 *eventData, NSString **result, NSError **error);
+int ZXPerformOcrThroughDaemon(NSString *payload, NSString **result, NSError **error);
 
 #endif